net: Fix up inet_addr_type checks
[ Upstream commit 30bbaa19500559d7625c65632195413f639b3b97 ]

Currently inet_addr_type and inet_dev_addr_type expect local addresses
to be in the local table. With the VRF device local routes for devices
associated with a VRF will be in the table associated with the VRF.
Provide an alternate inet_addr lookup to use a specific table rather
than defaulting to the local table.

inet_addr_type_dev_table keeps the same semantics as inet_addr_type but
if the passed in device is enslaved to a VRF then the table for that VRF
is used for the lookup.

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Lee: Back-ported to solve a build dependency]
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Change-Id: Ie93c7b4bcd98d9d4eca9558c87c24a4dba6c7b76