Revert "sched/core: Prevent race condition between cpuset and __sched_setscheduler()"
This reverts commit 710da3c8ea7dfbd327920afd3831d8c82c42789d.

When changing a thread's scheduling priority, binder calls
sched_setscheduler_nocheck() while holding the node lock and
proc inner lock. This was safe until v5.3 when this change was
introduced where cpuset_read_lock() is called in this path
which can sleep.

This patch was introduced to fix a possible accounting error
in sched deadline (potential oversell of CPU bandwidth) due
to a race condition between cpusets and __sched_setscheduler().
This is not an issue for Android.

This should be fixed in the binder driver, but that may take
some time.

[CPNOTE: 21/07/21] Lee: Requested status from the author via the bug
[CPNOTE: 07/09/21] Lee: Required whilst carrying current Binder PI implementation

Bug: 120446518 # OoT
Bug: 151861772
Change-Id: Ica1ef71b3cdcdc509b341ea1b57c41f8ee73794a
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Lee Jones <joneslee@google.com>