usb: gadget: f_hid: Conduct proper refcounting on shared f_hidg pointer
A pointer to the main HID gadget struct (*f_hidg) is shared with the
character device handing (read() and write(), etc support) on open().
However external references are presently not tracked.  This could
easily lead to some unsavoury behaviour if gadget support is disabled
/ disconnected.  Keeping track of the refcount ensures that resources
are only freed *after* they are no longer in use.

Signed-off-by: Lee Jones <lee@kernel.org>