usb: gadget: f_hid: Conduct proper refcounting on shared f_hidg pointer
A reference to struct f_hidg is shared with this driver's associated
character device handling component without provision for life-time
handling.  In some circumstances, this can lead to unwanted
behaviour depending on the order in which things are torn down.

Utilise, the reference counting functionality already provided by the
implanted character device structure to ensure the struct f_hidg's
memory is only freed once the character device handling has finished
with it.

Change-Id: I4b80a271b365f848ed3f256c8c8110a4c5c54752
Signed-off-by: Lee Jones <joneslee@google.com>