HID: core: Provide new max_buffer_size attribute to over-ride the default
Presently, when a report is processed, its size is compared solely
against the value specified by user-space.  If the received report ends
up being smaller than this, the remainder of the buffer is zeroed.  That
is, the space between sizeof(csize) (size of the current report) and the
rsize (size provided by the user - i.e. Report Size * Report Count),
which can be handled up to HID_MAX_BUFFER_SIZE (16k).

This is an issue.  In the case of some low-level drivers, the buffers
are significantly smaller than the default 16k.  For instance, in the
case of uhid, the data buffer is a mere UHID_DATA_MAX (4k).  Meaning
that, memset() shoots straight past the end of the buffer boundary and
starts zeroing out in-use values, often resulting in calamity.

This patch introduces a new variable into 'struct hid_ll_driver' where
individual low-level drivers can over-ride the default maximum value of
HID_MAX_BUFFER_SIZE (16k) with something more sympathetic to the
interface.

Signed-off-by: Lee Jones <lee@kernel.org>