BACKPORT: usb: gadget: uvc: Add struct for color matching in configs
Color matching descriptors are meant to be a per-format piece of data
and we need to be able to support different descriptors for different
formats. As a preliminary step towards that goal, switch the default
color matching configfs functionality to point to an instance of a
new struct uvcg_color_matching. Use the same default values for its
attributes as the currently hard-coded ones so that the interface to
userspace is consistent.

Bug: 254441685
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Link: https://lore.kernel.org/r/20230202114142.300858-3-dan.scally@ideasonboard.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 744eb7b821f61eceed60eb4f64227162853c9d5e)
[Lee: Dependency of 'usb: gadget: webcam: Make g_webcam loadable again']
Signed-off-by: Lee Jones <joneslee@google.com>
Change-Id: Ib7711a53e8d8b1b952d6604c7de90bf464ec302e