SMB3: Close deferred file handles if we get handle lease break
We should not cache deferred file handles if we dont have
handle lease on a file. And we should immediately close all
deferred handles in case of handle lease break. This is
very important e.g. to prevent access denied errors
in writing to the same file from a different client (during
this interval while the close is deferred).

Signed-off-by: Bharath SM <bharathsm@microsoft.com>
Fixes: 9e31678fb403 ("SMB3: fix lease break timeout when multiple deferred close handles for the same file.")
Cc: stable@vger.kernel.org # 6.0+
Signed-off-by: Steve French <stfrench@microsoft.com>