firmware: arm_ffa: Drop le64_to_cpu() on UUID in ffa_msg_send_direct_req2()
UUID is always saved in the big-endian format.
e.g: "378daedc-f06b-4446-8314-40ab933c87a3" will be saved in memory as:
    [0]
  |  37 8d ae dc
  |  f0 6b 44 46
  |  83 14 40 ab
  |  93 3c 87 a3
             [15]

As per the FF-A specification FFA_MSG_SEND_DRIECT_REQ2, UUID must be passed
in the registers as:

  |  UUID_Lo  x2  Bytes[0....7] of UUID with byte 0 in the low-order bits.
  |  UUID_Hi  x3  Bytes[8...15] of UUID with byte 8 in the low-order bits.

Commit 629253b2f6d7 ("firmware: arm_ffa: Avoid string-fortify warning in
export_uuid()") ended up swapping the bytes as the BE format was saved into
LE format/variable and then le64_to_cpu() resulted in a swap.

Drop the call to le64_to_cpu() for uuid in ffa_msg_send_direct_req2() and
change uuid_regs' type to unsigned long.

Currently the driver work in big-endian mode and needs more change if that
needs to be supported.

Fixes: 629253b2f6d7 ("firmware: arm_ffa: Avoid string-fortify warning in export_uuid()")
Signed-off-by: Levi Yun <yeoreum.yun@arm.com>
Message-Id: <20241203143109.1030514-3-yeoreum.yun@arm.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>